DefaultBCComms -> BCComms, keep Default.cpp for factory
