+++ CMakeLists.txt
@@ -77,21 +77,24 @@ if (WIN32 AND ${USE_CPPREST})
 	list(APPEND OS_SPECIFIC_SRCS "src/win/CppRestFileUploader.cpp")
 	list(APPEND OS_SPECIFIC_SRCS "src/win/MD5.cpp")
 elseif (USE_CURL)
-	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/DefaultBrainCloudComms.h")
+	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/BrainCloudComms.h")
+	list(APPEND OS_SPECIFIC_SRCS "src/BrainCloudComms.cpp")
 	list(APPEND OS_SPECIFIC_SRCS "src/DefaultBrainCloudComms.cpp")
 	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/nix/cURLFileUploader.h")
 	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/nix/cURLLoader.h")
 	list(APPEND OS_SPECIFIC_SRCS "src/nix/cURLFileUploader.cpp")
 	list(APPEND OS_SPECIFIC_SRCS "src/nix/cURLLoader.cpp")
 elseif (VS2013_OR_HIGHER)
-	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/DefaultBrainCloudComms.h")
+	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/BrainCloudComms.h")
+	list(APPEND OS_SPECIFIC_SRCS "src/BrainCloudComms.cpp")
 	list(APPEND OS_SPECIFIC_SRCS "src/DefaultBrainCloudComms.cpp")
 	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/win/XMLHTTPRequestFileUploader.h")
 	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/win/XMLHTTPRequestLoader.h")
 	list(APPEND OS_SPECIFIC_SRCS "src/win/XMLHTTPRequestFileUploader.cpp")
 	list(APPEND OS_SPECIFIC_SRCS "src/win/XMLHTTPRequestLoader.cpp")
 else()
-	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/DefaultBrainCloudComms.h")
+	list(APPEND OS_SPECIFIC_INCS "include/braincloud/internal/BrainCloudComms.h")
+	list(APPEND OS_SPECIFIC_SRCS "src/BrainCloudComms.cpp")
 	list(APPEND OS_SPECIFIC_SRCS "src/DefaultBrainCloudComms.cpp")
 endif()
